+++ src/opm/input/eclipse/Schedule/HandlerContext.cpp
@@ -25,9 +25,12 @@
 #include <opm/input/eclipse/Deck/DeckKeyword.hpp>
 
 #include <opm/input/eclipse/Parser/ParseContext.hpp>
+#include <opm/input/eclipse/Parser/ParserKeywords/W.hpp>
 
 #include <opm/input/eclipse/Schedule/Action/SimulatorUpdate.hpp>
 #include <opm/input/eclipse/Schedule/Schedule.hpp>
+#include <opm/input/eclipse/Schedule/Well/Well.hpp>
+
 
 #include "MSW/WelSegsSet.hpp"
 
@@ -166,6 +169,81 @@ void HandlerContext::addGroupToGroup(const std::string& parent_group,
                                      const std::string& child_group)
 {
     schedule_.addGroupToGroup(parent_group, child_group);
+}
+
+void HandlerContext::welspecsCreateNewWell(const DeckRecord&  record,
+                                           const std::string& wellName,
+                                           const std::string& groupName)
+{
+    auto wellConnectionOrder = Connection::Order::TRACK;
+
+    if (const auto& compord = block.get("COMPORD"); compord.has_value())
+    {
+        const auto nrec = compord->size();
+
+        for (auto compordRecordNr = 0*nrec; compordRecordNr < nrec; ++compordRecordNr) {
+            const auto& compordRecord = compord->getRecord(compordRecordNr);
+
+            const std::string& wellNamePattern = compordRecord.getItem(0).getTrimmedString(0);
+
+            if (Well::wellNameInWellNamePattern(wellName, wellNamePattern)) {
+                const std::string& compordString = compordRecord.getItem(1).getTrimmedString(0);
+                wellConnectionOrder = Connection::OrderFromString(compordString);
+            }
+        }
+    }
+
+    schedule_.addWell(wellName, record, currentStep, wellConnectionOrder);
+    schedule_.addWellToGroup(groupName, wellName, currentStep);
+
+    this->affected_well(wellName);
+}
+
+void HandlerContext::
+welspecsUpdateExistingWells(const DeckRecord&               record,
+                            const std::vector<std::string>& wellNames,
+                            const std::string&              groupName)
+{
+    using Kw = ParserKeywords::WELSPECS;
+
+    const auto& headI = record.getItem<Kw::HEAD_I>();
+    const auto& headJ = record.getItem<Kw::HEAD_J>();
+    const auto& pvt   = record.getItem<Kw::P_TABLE>();
+    const auto& drad  = record.getItem<Kw::D_RADIUS>();
+    const auto& ref_d = record.getItem<Kw::REF_DEPTH>();
+
+    const auto I = headI.defaultApplied(0) ? std::nullopt : std::optional<int> {headI.get<int>(0) - 1};
+    const auto J = headJ.defaultApplied(0) ? std::nullopt : std::optional<int> {headJ.get<int>(0) - 1};
+
+    const auto pvt_table = pvt.defaultApplied(0) ? std::nullopt : std::optional<int> { pvt.get<int>(0) };
+    const auto drainageRadius = drad.defaultApplied(0) ? std::nullopt : std::optional<double> { drad.getSIDouble(0) };
+
+    auto ref_depth = std::optional<double>{};
+    if (! ref_d.defaultApplied(0) && ref_d.hasValue(0)) {
+        ref_depth.emplace(ref_d.getSIDouble(0));
+    }
+
+    for (const auto& wellName : wellNames) {
+        auto well = state().wells.get(wellName);
+
+        const auto updateHead = well.updateHead(I, J);
+        const auto updateRefD = well.updateRefDepth(ref_depth);
+        const auto updateDRad = well.updateDrainageRadius(drainageRadius);
+        const auto updatePVT  = well.updatePVTTable(pvt_table);
+
+        if (updateHead || updateRefD || updateDRad || updatePVT) {
+            well.updateRefDepth();
+
+            state().wellgroup_events()
+                .addEvent(wellName, ScheduleEvents::WELL_WELSPECS_UPDATE);
+
+            state().wells.update(std::move(well));
+
+            this->affected_well(wellName);
+        }
+
+        this->schedule_.addWellToGroup(groupName, wellName, currentStep);
+    }
 }
 
 }
