+++ src/opm/input/eclipse/Schedule/Schedule.cpp
@@ -1756,83 +1756,6 @@ File {} line {}.)", pattern, location.keyword, location.filename, location.linen
         return sched_state->wlist_manager.get().hasList(pattern);
     }
 
-    void Schedule::welspecsCreateNewWell(const DeckRecord&  record,
-                                         const std::string& wellName,
-                                         const std::string& groupName,
-                                         HandlerContext&    handlerContext)
-    {
-        auto wellConnectionOrder = Connection::Order::TRACK;
-
-        if (const auto& compord = handlerContext.block.get("COMPORD");
-            compord.has_value())
-        {
-            const auto nrec = compord->size();
-
-            for (auto compordRecordNr = 0*nrec; compordRecordNr < nrec; ++compordRecordNr) {
-                const auto& compordRecord = compord->getRecord(compordRecordNr);
-
-                const std::string& wellNamePattern = compordRecord.getItem(0).getTrimmedString(0);
-
-                if (Well::wellNameInWellNamePattern(wellName, wellNamePattern)) {
-                    const std::string& compordString = compordRecord.getItem(1).getTrimmedString(0);
-                    wellConnectionOrder = Connection::OrderFromString(compordString);
-                }
-            }
-        }
-
-        this->addWell(wellName, record, handlerContext.currentStep, wellConnectionOrder);
-        this->addWellToGroup(groupName, wellName, handlerContext.currentStep);
-
-        handlerContext.affected_well(wellName);
-    }
-
-    void Schedule::welspecsUpdateExistingWells(const DeckRecord&               record,
-                                               const std::vector<std::string>& wellNames,
-                                               const std::string&              groupName,
-                                               HandlerContext&                 handlerContext)
-    {
-        using Kw = ParserKeywords::WELSPECS;
-
-        const auto& headI = record.getItem<Kw::HEAD_I>();
-        const auto& headJ = record.getItem<Kw::HEAD_J>();
-        const auto& pvt   = record.getItem<Kw::P_TABLE>();
-        const auto& drad  = record.getItem<Kw::D_RADIUS>();
-        const auto& ref_d = record.getItem<Kw::REF_DEPTH>();
-
-        const auto I = headI.defaultApplied(0) ? std::nullopt : std::optional<int> {headI.get<int>(0) - 1};
-        const auto J = headJ.defaultApplied(0) ? std::nullopt : std::optional<int> {headJ.get<int>(0) - 1};
-
-        const auto pvt_table = pvt.defaultApplied(0) ? std::nullopt : std::optional<int> { pvt.get<int>(0) };
-        const auto drainageRadius = drad.defaultApplied(0) ? std::nullopt : std::optional<double> { drad.getSIDouble(0) };
-
-        auto ref_depth = std::optional<double>{};
-        if (! ref_d.defaultApplied(0) && ref_d.hasValue(0)) {
-            ref_depth.emplace(ref_d.getSIDouble(0));
-        }
-
-        for (const auto& wellName : wellNames) {
-            auto well = this->snapshots.back().wells.get(wellName);
-
-            const auto updateHead = well.updateHead(I, J);
-            const auto updateRefD = well.updateRefDepth(ref_depth);
-            const auto updateDRad = well.updateDrainageRadius(drainageRadius);
-            const auto updatePVT  = well.updatePVTTable(pvt_table);
-
-            if (updateHead || updateRefD || updateDRad || updatePVT) {
-                well.updateRefDepth();
-
-                this->snapshots.back().wellgroup_events()
-                    .addEvent(wellName, ScheduleEvents::WELL_WELSPECS_UPDATE);
-
-                this->snapshots.back().wells.update(std::move(well));
-
-                handlerContext.affected_well(wellName);
-            }
-
-            this->addWellToGroup(groupName, wellName, handlerContext.currentStep);
-        }
-    }
-
     const std::map< std::string, int >& Schedule::rst_keywords( size_t report_step ) const {
         if (report_step == 0)
             return this->m_static.rst_config.keywords;
